+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file including pl
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.2] - 2023-04-03
+
+### Added 
+- Result export button
+
+### Fixed
+- Function Button is translatable
+- Aim is translatable
 
 ## [0.1.1] - 2023-03-28
 
@@ -32,5 +40,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Hint
 - ListBox
 
+[0.1.1]: https://github.com/tblanke/ScenarioGUI/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/tblanke/ScenarioGUI/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/tblanke/ScenarioGUI/releases/tag/v0.1.0